Cleaned up a few more instances of counters that should be converted to uint64_t in the data ingestion pipeline. (#2405)

+++ include/lbann/callbacks/variable_minibatch.hpp
@@ -42,7 +42,7 @@ namespace callback {
 class variable_minibatch : public callback_base
 {
 public:
-  variable_minibatch(size_t starting_mbsize);
+  variable_minibatch(uint64_t starting_mbsize);
   variable_minibatch(const variable_minibatch&) = default;
   variable_minibatch& operator=(const variable_minibatch&) = default;
   /// Set the initial mini-batch size.
